VSMT_UPROT_APP/requirements.txt
flask
requests
fhir.resources
pymongo
flask-session
redis